Clean up all READMEs (#1)
* Clean up all READMEs

My goal with this original pass through of the READMEs is to:

* Fix a variety of grammatical and spelling mistakes
* Improve the readibility of various sections
* Ensure links work correctly
* Make the adventures more consistent with one another

There are still many issues that I didn't address. For instance,
some of the adventures did not work for me when I ran through them
and some of the instructions linked to things that didn't appear to
exist.

I also didn't improve the steps themselves to break things down
further. More work will be needed in the future, but I wanted to
save these changes for now.

* Fix typo in main README

* Consistent spring-petclinic

* Use relative links

---------

Co-authored-by: Mike Sol <mike-solomon@users.noreply.github.com>
Co-authored-by: Tim te Beek <tim@moderne.io>